// src/interfaces/interfaces.ts
export interface QuestionTopic {
  id: number;
  title: string;
}

//report coming from the external world
export interface Report {
  client_id: string;
  question_id: number;
  reason: string;
}

//report handled by the app
export interface ReportHandled {
  username: string;
  question_title: string;
  timestamp: Date;
  topic_id: number;
  topic_title: string;
  question_id: number;
  reason: string;
}

export interface Question {
  id: number;
  title: string;
  timestamp: Date;
  topic: QuestionTopic;
  n?: number;
}
// ...
export enum Lang {
  IT = "it",
  EN = "en",
  FR = "fr",
  ES = "es",
}
// ...
export enum TopicLevel {
  EASY,
  MEDIUM,
  HARD,
}